Time: 1 ms (98.80%), Space: 39.3 MB (92.78%) - LeetHub

+++ 69-sqrtx/69-sqrtx.java
@@ -1,7 +1,7 @@
 import java.lang.Math;
 class Solution {
     public int mySqrt(int x) {
-        double a=Math.sqrt(x);
-        return (int)a;
+        int a=(int)Math.sqrt(x);
+        return a;
     }
 }
